4.1.1 Correcciones de estilo en tareas update

+++ ofw/task/update.php
@@ -17,19 +17,19 @@ class updateTask{
       }
     }
     asort($to_be_updated);
-    echo "Se han encontrado ".count($to_be_updated)." actualizaciones pendientes. Se procede a la instalación ordenada.\n\n";
+    echo "  Se han encontrado ".count($to_be_updated)." actualizaciones pendientes. Se procede a la instalación ordenada.\n\n";
 
     foreach ($to_be_updated as $repo_version){
       $backups = [];
       $result = true;
-      echo "  ".$repo_version." : ".$updates[$repo_version]['message']."\n";
-      echo "==============================================================================================================\n\n";
+      echo "  ".$updates[$repo_version]['message']."\n";
+      echo "==============================================================================================================\n";
 
-      if (count($updates[$repo_version]['deletes'])>0){
+      if (array_key_exists('deletes', $updates[$repo_version]) && count($updates[$repo_version]['deletes'])>0){
         foreach ($updates[$repo_version]['deletes'] as $delete){
           $local_delete = $c->getDir('base').$delete;
           if (file_exists($local_delete)){
-            echo " El archivo \"".$local_delete."\" será eliminado.\n";
+            echo " El archivo \"".$delete."\" será eliminado.\n";
             $backup_file = $local_delete.'_backup';
             rename($local_delete, $backup_file);
             array_push($backups, ['new_file'=>$local_delete, 'backup'=>$backup_file]);
@@ -37,21 +37,21 @@ class updateTask{
         }
         echo "\n";
       }
-      if (count($updates[$repo_version]['files'])>0){
+      if (array_key_exists('files', $updates[$repo_version]) && count($updates[$repo_version]['files'])>0){
         foreach ($updates[$repo_version]['files'] as $file){
           $file_url = $this->repo_url.$file;
-          echo "Descargando \"".$file_url."\"\n";
+          echo "  Descargando \"".$file_url."\"\n";
           $file_content = file_get_contents($file_url);
 
           $local_file = $c->getDir('base').$file;
           if (file_exists($local_file)){
-            echo "  El archivo ya existe, creando copia de seguridad.\n";
+            echo "    El archivo ya existe, creando copia de seguridad.\n";
             $backup_file = $local_file.'_backup';
             rename($local_file, $backup_file);
             array_push($backups, ['new_file'=>$local_file, 'backup'=>$backup_file]);
           }
           else{
-            echo "Creando nuevo archivo.\n";
+            echo "    Creando nuevo archivo.\n";
           }
 
           $result_file = file_put_contents($local_file, $file_content);
@@ -61,20 +61,21 @@ class updateTask{
           }
         }
       }
+      echo "==============================================================================================================\n";
 
       if ($result){
         $version_file = $c->getDir('ofw_base').'VERSION';
         file_put_contents($version_file, $repo_version);
-        echo "\nTodos los archivos han sido actualizados. La nueva versión instalada es: ".$repo_version."\n";
+        echo "\n  Todos los archivos han sido actualizados. La nueva versión instalada es: ".$repo_version."\n";
         if (count($backups)>0){
-          echo "Se procede a eliminar las copias de seguridad realizadas.\n";
+          echo "  Se procede a eliminar las copias de seguridad realizadas.\n";
           foreach ($backups as $backup){
             unlink($backup['backup']);
           }
         }
       }
       else{
-        echo "Ocurrió un error al actualizar los archivos, se procede a restaurar las copias de seguridad.\n";
+        echo "  Ocurrió un error al actualizar los archivos, se procede a restaurar las copias de seguridad.\n";
         foreach ($backups as $backup){
           if (file_exists($backup['new_file'])){
             unlink($backup['new_file']);
@@ -82,6 +83,7 @@ class updateTask{
           rename($backup['backup'], $backup['new_file']);
         }
       }
+      echo "\n";
     }
   }
 
@@ -89,23 +91,25 @@ class updateTask{
     $current_version = trim( Base::getVersion() );
     $repo_version = trim( file_get_contents($this->repo_url.'ofw/base/VERSION') );
 
-    echo "Versión instalada: ".$current_version."\n";
-    echo "Versión actual: ".$repo_version."\n";
+    echo "\n";
+    echo "  Osumi Framework\n";
+    echo "  Versión instalada: ".$current_version."\n";
+    echo "  Versión actual: ".$repo_version."\n";
 
     $compare = version_compare($current_version, $repo_version);
 
     switch ($compare){
       case -1: {
-        echo "Se procede a la actualización.\n";
+        echo "  Se procede a la actualización.\n";
         $this->doUpdate($current_version);
       }
       break;
       case 0: {
-        echo "La versión instalada está actualizada.\n";
+        echo "  La versión instalada está actualizada.\n";
       }
       break;
       case 1: {
-        echo "¡¡La versión instalada está MÁS actualizada que la del repositorio!!\n";
+        echo "  ¡¡La versión instalada está MÁS actualizada que la del repositorio!!\n";
       }
       break;
     }
